% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{atomic_save_rds}
\alias{atomic_save_rds}
\title{Atomic RDS writer}
\usage{
atomic_save_rds(object, path)
}
\arguments{
\item{object}{Object to save.}

\item{path}{Destination.}
}
\description{
Atomic RDS writer
}
