% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_backtesting.R
\name{backtest_var}
\alias{backtest_var}
\title{Complete VaR backtest summary}
\usage{
backtest_var(realised_returns, var, confidence = 0.99)
}
\arguments{
\item{realised_returns}{Signed returns.}

\item{var}{Positive VaR.}

\item{confidence}{Confidence level.}
}
\description{
Complete VaR backtest summary
}
