% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_returns.R
\name{calculate_portfolio_returns}
\alias{calculate_portfolio_returns}
\title{Calculate exact fixed-weight portfolio returns}
\usage{
calculate_portfolio_returns(
  asset_simple_returns,
  weights = NULL,
  allow_short = FALSE
)
}
\arguments{
\item{asset_simple_returns}{Matrix of asset simple returns.}

\item{weights}{Portfolio weights.}

\item{allow_short}{Whether shorts are allowed.}
}
\value{
Data frame with portfolio simple and exact log returns.
}
\description{
Calculate exact fixed-weight portfolio returns
}
