% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_backtesting.R
\name{christoffersen_conditional_test}
\alias{christoffersen_conditional_test}
\title{Christoffersen conditional coverage test}
\usage{
christoffersen_conditional_test(exceptions, confidence = 0.99)
}
\arguments{
\item{exceptions}{Logical indicators.}

\item{confidence}{VaR confidence level.}
}
\value{
Named test-result list.
}
\description{
Christoffersen conditional coverage test
}
