% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_backtesting.R
\name{christoffersen_independence_test}
\alias{christoffersen_independence_test}
\title{Christoffersen independence test}
\usage{
christoffersen_independence_test(exceptions)
}
\arguments{
\item{exceptions}{Logical exception indicators.}
}
\value{
Named test-result list.
}
\description{
Christoffersen independence test
}
