% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{common_score_dates}
\alias{common_score_dates}
\title{Find dates with a finite score for every model}
\usage{
common_score_dates(x, score_column)
}
\description{
Find dates with a finite score for every model
}
\keyword{internal}
