% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail_dependence.R
\name{copula_tail_dependence}
\alias{copula_tail_dependence}
\title{Analytical tail dependence for native copulas}
\usage{
copula_tail_dependence(family, parameters)
}
\arguments{
\item{family}{Copula family.}

\item{parameters}{Parameter list.}
}
\value{
Named lower and upper coefficients.
}
\description{
Analytical tail dependence for native copulas
}
