% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_backtesting.R
\name{coverage_traffic_light}
\alias{coverage_traffic_light}
\title{Traffic-light coverage label}
\usage{
coverage_traffic_light(p_value, exception_rate, expected_rate)
}
\arguments{
\item{p_value}{Conditional-coverage p-value.}

\item{exception_rate}{Observed rate.}

\item{expected_rate}{Expected rate.}
}
\description{
Traffic-light coverage label
}
