% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ingestion.R
\name{data_request_signature}
\alias{data_request_signature}
\title{Build a complete data-request signature}
\usage{
data_request_signature(cfg, root)
}
\description{
Build a complete data-request signature
}
\keyword{internal}
