% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail_dependence.R
\name{empirical_tail_dependence}
\alias{empirical_tail_dependence}
\title{Empirical lower and upper tail dependence estimates}
\usage{
empirical_tail_dependence(u, threshold = 0.05)
}
\arguments{
\item{u}{Two-column matrix of uniform observations.}

\item{threshold}{Tail probability threshold.}
}
\value{
Named list of threshold estimates.
}
\description{
Empirical lower and upper tail dependence estimates
}
