% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_forecast.R
\name{empty_rolling_forecasts}
\alias{empty_rolling_forecasts}
\title{Empty common-schema rolling result}
\usage{
empty_rolling_forecasts()
}
\description{
Empty common-schema rolling result
}
\keyword{internal}
