% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_project_dirs}
\alias{ensure_project_dirs}
\title{Create all project runtime directories}
\usage{
ensure_project_dirs(root = pra_project_root())
}
\arguments{
\item{root}{Project root.}
}
\value{
Named character vector of directories.
}
\description{
Create all project runtime directories
}
