% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_metrics.R
\name{estimate_t_df}
\alias{estimate_t_df}
\title{Estimate Student-t degrees of freedom by profile likelihood}
\usage{
estimate_t_df(standardised_returns, bounds = c(2.05, 100))
}
\arguments{
\item{standardised_returns}{Approximately standardised observations.}

\item{bounds}{Search bounds, both above two.}
}
\value{
Estimated degrees of freedom.
}
\description{
Estimate Student-t degrees of freedom by profile likelihood
}
