% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_backtesting.R
\name{exception_transitions}
\alias{exception_transitions}
\title{Christoffersen transition counts}
\usage{
exception_transitions(exceptions)
}
\arguments{
\item{exceptions}{Logical indicators.}
}
\value{
Named integer vector n00, n01, n10, n11.
}
\description{
Christoffersen transition counts
}
