% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_models.R
\name{fit_rugarch_marginal}
\alias{fit_rugarch_marginal}
\title{Fit a rugarch marginal model}
\usage{
fit_rugarch_marginal(returns, spec, solver = "hybrid")
}
\description{
Fit a rugarch marginal model
}
\keyword{internal}
