% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_models.R
\name{gaussian_copula_loglik}
\alias{gaussian_copula_loglik}
\title{Bivariate Gaussian copula log-likelihood}
\usage{
gaussian_copula_loglik(u, rho)
}
\description{
Bivariate Gaussian copula log-likelihood
}
\keyword{internal}
