% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{implementation_fingerprint}
\alias{implementation_fingerprint}
\title{Identify the implementation and numerical dependency versions}
\usage{
implementation_fingerprint()
}
\description{
Identify the implementation and numerical dependency versions
}
\keyword{internal}
