% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{make_run_id}
\alias{make_run_id}
\title{Build a deterministic run identifier}
\usage{
make_run_id(profile, when = Sys.time())
}
\arguments{
\item{profile}{Profile name.}

\item{when}{Timestamp.}
}
\description{
Build a deterministic run identifier
}
