% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_asset_prices}
\alias{plot_asset_prices}
\title{Plot aligned asset prices}
\usage{
plot_asset_prices(aligned_prices, path)
}
\arguments{
\item{aligned_prices}{Wide price data.}

\item{path}{Output PNG path.}
}
\description{
Plot aligned asset prices
}
