% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_copula_diagnostic}
\alias{plot_copula_diagnostic}
\title{Plot PIT dependence}
\usage{
plot_copula_diagnostic(u, simulated, path)
}
\arguments{
\item{u}{Observed PIT matrix.}

\item{simulated}{Simulated copula values.}

\item{path}{Output path.}
}
\description{
Plot PIT dependence
}
