% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_model_comparison}
\alias{plot_model_comparison}
\title{Plot model comparison quantile loss}
\usage{
plot_model_comparison(comparison, path)
}
\arguments{
\item{comparison}{Model comparison table.}

\item{path}{Output path.}
}
\description{
Plot model comparison quantile loss
}
