% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_monte_carlo_convergence}
\alias{plot_monte_carlo_convergence}
\title{Plot Monte Carlo convergence}
\usage{
plot_monte_carlo_convergence(summary, path)
}
\arguments{
\item{summary}{Convergence summary table.}

\item{path}{Output path.}
}
\description{
Plot Monte Carlo convergence
}
