% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_simulation_risk}
\alias{plot_simulation_risk}
\title{Plot portfolio simulation and risk thresholds}
\usage{
plot_simulation_risk(simulation, path)
}
\arguments{
\item{simulation}{Simulation result.}

\item{path}{Output path.}
}
\description{
Plot portfolio simulation and risk thresholds
}
