% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pra_project_root}
\alias{pra_project_root}
\title{Locate the project root}
\usage{
pra_project_root(start = getwd())
}
\arguments{
\item{start}{Directory from which to search.}
}
\value{
Normalised project-root path.
}
\description{
Walks upward until \code{DESCRIPTION} is found. No absolute path is stored.
}
