% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{project_path}
\alias{project_path}
\title{Construct a project-relative path}
\usage{
project_path(..., root = pra_project_root())
}
\arguments{
\item{...}{Path components below the project root.}

\item{root}{Project root.}
}
\description{
Construct a project-relative path
}
