% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{source_revision}
\alias{source_revision}
\title{Capture the source revision without depending on Git being installed}
\usage{
source_revision(root)
}
\description{
Capture the source revision without depending on Git being installed
}
\keyword{internal}
