% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_models.R
\name{student_copula_loglik}
\alias{student_copula_loglik}
\title{Bivariate Student-t copula log-likelihood}
\usage{
student_copula_loglik(u, rho, df)
}
\description{
Bivariate Student-t copula log-likelihood
}
\keyword{internal}
