% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_stage}
\alias{time_stage}
\title{Time an expression}
\usage{
time_stage(expr)
}
\arguments{
\item{expr}{Expression to evaluate.}
}
\value{
List with value and elapsed seconds.
}
\description{
Time an expression
}
