% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{uniform_ad_statistic}
\alias{uniform_ad_statistic}
\title{Anderson-Darling statistic for uniformity}
\usage{
uniform_ad_statistic(u)
}
\arguments{
\item{u}{Values in (0,1).}
}
\value{
Finite test statistic.
}
\description{
Anderson-Darling statistic for uniformity
}
