% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_config}
\alias{validate_config}
\title{Validate analysis configuration}
\usage{
validate_config(cfg)
}
\arguments{
\item{cfg}{Configuration list.}
}
\value{
Configuration, invisibly.
}
\description{
Validate analysis configuration
}
