% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_backtesting.R
\name{var_exceptions}
\alias{var_exceptions}
\title{Determine VaR exceedances}
\usage{
var_exceptions(realised_returns, var)
}
\arguments{
\item{realised_returns}{Signed realised returns.}

\item{var}{Positive VaR magnitudes.}
}
\value{
Logical vector, \code{-return > VaR}.
}
\description{
Determine VaR exceedances
}
