% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_models.R
\name{vine_family_code}
\alias{vine_family_code}
\title{Map readable names to VineCopula family codes}
\usage{
vine_family_code(family)
}
\description{
Map readable names to VineCopula family codes
}
\keyword{internal}
