% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{write_png}
\alias{write_png}
\title{Open a project PNG device and evaluate plotting code}
\usage{
write_png(path, code, width = 1400, height = 850, res = 140)
}
\description{
Open a project PNG device and evaluate plotting code
}
\keyword{internal}
