% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{compare_forecast_models}
\alias{compare_forecast_models}
\title{Compare rolling forecast models}
\usage{
compare_forecast_models(forecasts)
}
\arguments{
\item{forecasts}{Common-schema rolling forecasts.}
}
\value{
Model comparison table.
}
\description{
Compare rolling forecast models
}
