% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnose_marginal}
\alias{diagnose_marginal}
\title{Residual and PIT diagnostics}
\usage{
diagnose_marginal(fit, lag = 10L, pit_epsilon = 1e-08)
}
\arguments{
\item{fit}{Successful marginal fit.}

\item{lag}{Ljung-Box lag.}

\item{pit_epsilon}{PIT clipping value.}
}
\value{
Named diagnostic list.
}
\description{
Residual and PIT diagnostics
}
